requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.8.0
einops>=0.7.0
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
mamba-ssm>=1.2.0 ; platform_system == "Linux"
causal-conv1d>=1.2.0 ; platform_system == "Linux"
torch-geometric>=2.5.0
timm>=0.9.0
sentencepiece>=0.2.0
accelerate>=0.27.0
